Extracts ChangeEmitter.js - Refactors EventStore.js
